requirements.txt
langchain
chromadb
pypdf 
python-dotenv
streamlit